fixed hopfully last error

+++ cfm/cfm-code/EmModel.cpp
@@ -678,7 +678,7 @@ double EmModel::gaGetUpdatedLearningRate(double learning_rate, int iter) const {
   return learning_rate;
 }
 
-int EmModel::computeAndAccumulateGradient(double *grads, int mol_idx,
+int EmModel::computeAndAccumulateGradient(float *grads, int mol_idx,
                                           MolData &mol_data,
                                           suft_counts_t &suft,
                                           int sampling_method,
