+++ cfm/cfm-code/EmNNModel.h
@@ -36,9 +36,8 @@ public:
             std::string initial_params_filename = "");
 
   // This is public so the test can access it....there must be a better way?
-  int computeAndAccumulateGradient(double *grads, int mol_idx,
-                                   MolData &mol_data, suft_counts_t &suft,
-                                   int sampling_method,
+  int computeAndAccumulateGradient(float *grads, int mol_idx, MolData &mol_data,
+                                   suft_counts_t &suft, int sampling_method,
                                    unsigned int energy) override;
 
   void collectUsedIdx(MolData &mol_data, std::set<unsigned int> &used_idxs,
